Add Calc game

+++ app/src/main/java/hexlet/code/Cli.java
@@ -12,12 +12,12 @@ public class Cli {
         return name;
     }
 
-    public static short promptGame(String[] games, final short exitIdx) {
+    public static int promptGame(String[] games, final int exitIdx) {
         String exitValue = games[exitIdx];
-        short gamesCount = (short) games.length;
+        int gamesCount = (int) games.length;
 
         System.out.println("Please enter the game number and press Enter.");
-        for (short i = 0; i < gamesCount; i += 1) {
+        for (int i = 0; i < gamesCount; i += 1) {
             if (exitIdx == i) {
                 continue;
             }
@@ -26,7 +26,7 @@ public class Cli {
         System.out.println(exitIdx + " - " + exitValue);
 
         System.out.print("Your choice: ");
-        short gameNumber = scanner.nextShort();
+        int gameNumber = scanner.nextShort();
         System.out.println();
 
         return gameNumber;
